Create DAY-13-Add two numbers by taking a password
The following program asks the user to create a password and then it takes the two numbers as input from the user in order to perform addition on them. But before displaying the result of addition, it again asks the user to enter the password which we had given earlier, and if the password entered is correct then it will display the result of the addition, but if the password entered is incorrect then it shows that the password you entered is wrong.